libnxsl, libnxdb, libnxsrv, nxget, and nxdbmgr are now UNICODE compatible; libtre intergrated for regexp support

+++ src/libnxlp/file.cpp
@@ -59,7 +59,7 @@ static void ParseNewRecords(LogParser *parser, int fh)
          bytes += bufPos;
          for(ptr = buffer;; ptr = eptr + 1)
          {
-            bufPos = ptr - buffer;
+            bufPos = (int)(ptr - buffer);
             eptr = (char *)memchr(ptr, '\n', bytes - bufPos);
             if (eptr == NULL)
             {
